requirements.txt
absl-py==2.0.0
accelerate==0.25.0
addict==2.4.0
aiohttp==3.9.1
aiosignal==1.3.1
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.42.0
boto3==1.34.32
botocore==1.34.32
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
datasets==2.16.1
diffusers==0.14.0
dill==0.3.7
Django==5.0.1
djangorestframework==3.14.0
dlib==19.24.2
docker-pycreds==0.4.0
drf-extra-fields==3.7.0
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.1.3
gitdb==4.0.11
GitPython==3.1.40
grpcio==1.60.0
huggingface-hub==0.20.1
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
inquirerpy==0.3.4
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.5
lazy_loader==0.3
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
mediapipe==0.10.9
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2.1
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.1.4
pfzy==0.3.4
Pillow==10.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.7
pyarrow==14.0.2
pyarrow-hotfix==0.6
pycocotools==2.0.7
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
s3transfer==0.10.0
safetensors==0.4.1
scikit-image==0.22.0
scipy==1.11.4
segment-anything==1.0
sentry-sdk==1.39.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sounddevice==0.4.6
sqlparse==0.4.4
supervision==0.17.1
sympy==1.12
tensorboard==1.15.0
tifffile==2023.12.9
timm==0.9.12
tokenizers==0.15.0
tomli==2.0.1
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.36.2
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
wandb==0.16.1
wcwidth==0.2.12
Werkzeug==3.0.1
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
